Chatbot (lex aws)/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
boto3==1.24.0
python-dotenv==1.0.0
